feat: consolidate cache logic into abstract class
