+++ core/src/main/kotlin/org/snappy/cache/DecoderCache.kt
@@ -1,60 +1,25 @@
 @file:Suppress("UNCHECKED_CAST")
-package org.snappy.decode
+package org.snappy.cache
 
-import io.github.classgraph.ClassGraph
 import io.github.classgraph.ScanResult
 import org.snappy.NoDefaultConstructor
 import org.snappy.SnappyConfig
 import org.snappy.TypeArgumentMismatch
+import org.snappy.decode.Decoder
 import org.snappy.logging.logger
-import java.util.concurrent.ConcurrentHashMap
 import kotlin.reflect.KClass
 import kotlin.reflect.KType
 import kotlin.reflect.full.createInstance
 import kotlin.reflect.full.createType
 import kotlin.reflect.full.isSubclassOf
-import kotlin.reflect.full.withNullability
 import kotlin.reflect.jvm.jvmErasure
 import kotlin.reflect.typeOf
 
-class DecoderCache internal constructor(private val config: SnappyConfig) {
+class DecoderCache internal constructor(
+    config: SnappyConfig,
+): AbstractTypeCache<Decoder<*>>(config) {
     /** [KLogger][io.github.oshai.kotlinlogging.KLogger] for this cache instance */
-    private val log by logger()
-    /** Internally used flag to indicate if the cache has been loaded */
-    internal var cacheLoaded: Boolean = false
-        private set
-    /** Map of an output type linked to a [Decoder] */
-    internal val cache: ConcurrentHashMap<KType, Decoder<*>> by lazy { initCache() }
-
-    private fun initCache(): ConcurrentHashMap<KType, Decoder<*>> {
-        val start = System.currentTimeMillis()
-        val cache = ConcurrentHashMap<KType, Decoder<*>>()
-        val packages = config.packages.toTypedArray()
-        ClassGraph().enableAllInfo().acceptPackages(*packages).scan().use { result ->
-            for((type, decoder) in processAllAutoCacheClasses(result)) {
-                if (cache.contains(type)) {
-                    log.warn {
-                        "Multiple decoders exist for type '$type'. The latest one seen is kept"
-                    }
-                }
-                cache[type] = decoder
-            }
-        }
-        cacheLoaded = true
-        val end = System.currentTimeMillis()
-        log.info {
-            "Decoder cache has been initialized. Took %.2f seconds".format((end - start)/1000.0)
-        }
-        return cache
-    }
-
-    /**
-     * Get a [Decoder] for the provided type [rowType]. Checks the [cache] for an existing
-     * parser and returns immediately if it exists. Otherwise, it returns null
-     */
-    fun getOrNull(rowType: KType): Decoder<*>? {
-        return cache[rowType]
-    }
+    override val log by logger()
 
     /**
      * Get a [Decoder] for the provided type [T]. Checks the [cache] for an existing
@@ -64,22 +29,13 @@ class DecoderCache internal constructor(private val config: SnappyConfig) {
         return getOrNull(typeOf<T>()) as Decoder<T>?
     }
 
-    /**
-     * Method to ensure the cache is loaded before continuing. This will force the lazy initialized
-     * caches to be loaded immediately in a blocking but thread-safe manner. This reduces the first
-     * load time of queries within the application.
-     */
-    internal fun loadCache() {
-        cache
-    }
-
-    private val decoderInterfaceKClass = Decoder::class
+    override val cacheType = Decoder::class
 
     /**
      * Yield pairs of a [KType] and [Decoder] to initialize the cache with classes implementing
      * [Decoder]
      */
-    private fun processAllAutoCacheClasses(result: ScanResult) = sequence {
+    override fun processAllAutoCacheClasses(result: ScanResult) = sequence {
         for (classInfo in result.getClassesImplementing(Decoder::class.java)) {
             if (
                 classInfo.isAbstract ||
@@ -99,7 +55,7 @@ class DecoderCache internal constructor(private val config: SnappyConfig) {
      */
     private fun parseDecoder(kClass: KClass<Decoder<*>>): Pair<KType, Decoder<*>> {
         val valueClassTypeArgument = kClass.supertypes
-            .first { it.jvmErasure.isSubclassOf(decoderInterfaceKClass) }
+            .first { it.jvmErasure.isSubclassOf(cacheType) }
             .arguments
             .first()
             .type!!
