+++ core/src/main/kotlin/org/snappy/ksp/symbols/RowParser.kt
@@ -1,4 +1,4 @@
-package org.snappy.ksp
+package org.snappy.ksp.symbols
 
 @Target(AnnotationTarget.CLASS)
 annotation class RowParser
