+++ ksp-test/src/main/kotlin/org/snappy/ksp/DataClassRowParserTest.kt
@@ -1,6 +1,7 @@
 package org.snappy.ksp
 
-import org.snappy.annotations.Rename
+import org.snappy.ksp.symbols.Rename
+import org.snappy.ksp.symbols.RowParser
 
 @RowParser
 data class DataClassRowParserTest(@Rename("name") val test: String)
