+++ ksp-test/src/main/kotlin/org/snappy/ksp/PojoClassRowParserTest.kt
@@ -1,5 +1,7 @@
 package org.snappy.ksp
 
+import org.snappy.ksp.symbols.RowParser
+
 @RowParser
 class PojoClassRowParserTest {
     var test: String = ""
