+++ ksp-test/src/test/kotlin/org/snappy/data/ResultPair.kt
@@ -1,6 +1,6 @@
 package org.snappy.data
 
-import org.snappy.ksp.RowParser
+import org.snappy.ksp.symbols.RowParser
 
 @RowParser
 data class ResultPair(val first: Int, val second: String)
